+++ cares_reinforcement_learning/networks/TD3/__init__.py
@@ -1,4 +1,2 @@
 from .actor import Actor
 from .critic import Critic
-from .cnn_actor import CNNActor
-from .cnn_critic import CNNCritic
